use size_t

+++ src/Open3D/Container/Kernel/UnaryEWKernel.cu
@@ -31,19 +31,19 @@
 
 #include "Open3D/Container/CudaUtils.h"
 
-static constexpr int threads_per_block = 128;
-static constexpr int items_per_thread = 4;
-constexpr int MAX_DIMS = 25;
+static constexpr size_t threads_per_block = 128;
+static constexpr size_t items_per_thread = 4;
+constexpr size_t MAX_DIMS = 25;
 
 namespace open3d {
 namespace kernel {
 
-template <int threads_per_block, int items_per_thread, typename func_t>
-__global__ void elementwise_kernel(int N, func_t f) {
-    int items_per_block = threads_per_block * items_per_thread;
-    int idx = blockIdx.x * items_per_block + threadIdx.x;
+template <size_t threads_per_block, size_t items_per_thread, typename func_t>
+__global__ void elementwise_kernel(size_t N, func_t f) {
+    size_t items_per_block = threads_per_block * items_per_thread;
+    size_t idx = blockIdx.x * items_per_block + threadIdx.x;
 #pragma unroll
-    for (int i = 0; i < items_per_thread; i++) {
+    for (size_t i = 0; i < items_per_thread; i++) {
         if (idx < N) {
             f(idx);
             idx += threads_per_block;
@@ -57,35 +57,35 @@ OPEN3D_HOST_DEVICE void templated_copy(const void* src, void* dst) {
 }
 
 struct OffsetCalculator {
-    OffsetCalculator(int num_dims,
+    OffsetCalculator(size_t num_dims,
                      const size_t* src_strides,
                      const size_t* dst_strides)
         : num_dims_(num_dims) {
-        for (int i = 0; i < num_dims_; i++) {
+        for (size_t i = 0; i < num_dims_; i++) {
             src_strides_[i] = src_strides[i];
             dst_strides_[i] = dst_strides[i];
         }
     }
 
-    OPEN3D_HOST_DEVICE int GetOffset(int idx) const {
-        int src_idx = 0;
-        for (int dim = 0; dim < num_dims_; dim++) {
+    OPEN3D_HOST_DEVICE size_t GetOffset(size_t idx) const {
+        size_t src_idx = 0;
+        for (size_t dim = 0; dim < num_dims_; dim++) {
             src_idx += idx / dst_strides_[dim] * src_strides_[dim];
             idx = idx % dst_strides_[dim];
         }
         return src_idx;
     }
 
-    int num_dims_;
-    int src_strides_[MAX_DIMS];
-    int dst_strides_[MAX_DIMS];
+    size_t num_dims_;
+    size_t src_strides_[MAX_DIMS];
+    size_t dst_strides_[MAX_DIMS];
 };
 
 template <typename T>
 static void CopyToContiguousCUDASameDevice(const Tensor& src, Tensor& dst) {
-    int N = static_cast<int>(src.GetShape().NumElements());
-    int items_per_block = threads_per_block * items_per_thread;
-    int grid_size = (N + items_per_block - 1) / items_per_block;
+    size_t N = static_cast<size_t>(src.GetShape().NumElements());
+    size_t items_per_block = threads_per_block * items_per_thread;
+    size_t grid_size = (N + items_per_block - 1) / items_per_block;
 
     const uint8_t* src_data_ptr = static_cast<const uint8_t*>(src.GetDataPtr());
     uint8_t* dst_data_ptr = static_cast<uint8_t*>(dst.GetDataPtr());
@@ -94,8 +94,8 @@ static void CopyToContiguousCUDASameDevice(const Tensor& src, Tensor& dst) {
                                        src.GetStrides().data(),
                                        dst.GetStrides().data());
 
-    auto f = [=] OPEN3D_HOST_DEVICE(int idx) {
-        int src_idx = offset_calculator.GetOffset(idx);
+    auto f = [=] OPEN3D_HOST_DEVICE(size_t idx) {
+        size_t src_idx = offset_calculator.GetOffset(idx);
         const void* src_ptr = src_data_ptr + src_idx * element_byte_size;
         void* dst_ptr = dst_data_ptr + idx * element_byte_size;
         templated_copy<T>(src_ptr, dst_ptr);
